files

+++ 0x00-python_variable_annotations/3-to_str.py
@@ -4,7 +4,7 @@
 '''
 
 
-def to_str(n: int) -> str:
+def to_str(n: float) -> str:
     '''function to_str that takes a float n as
     argument and returns the string representation of the float.
     '''
